requirements.txt
mcp>=0.9.0
openai>=1.0.0
python-dotenv>=1.0.0